Replace simple filters by Treat parser
The treat parser can parse a simple filter query language which consists
of treats in a form of [key][operator][value], e.g. extension:pdf and
arbitary text which is considered to be the actual search query.

The prior approach of adding little checkboxes is completely abandoned,
please use the Treat parser instead.

Treats are interpreted as AND filters for elasticsearch. OR filters are
currently not possible and have to be done by doing multiple search
queries. Adding a filter multiple times refines the search result but
might be order dependend.

Currently working treats are:
* extension:[fileextension] and extension![fileextension]
* size>[humanized filesize] and size<[humanized filesize]

+ Make the large file booster obligatory
+ Remove the small file filter as the large file booster should do the
  same job while not hiding any search results

TODO add frontend documentation on how to use treats
TODO add "we interpreted your query like […]" to the frontend
TODO add more developers documentation about the treats